+++ packages/dnb-eufemia-sandbox/.storybook/main.js
@@ -1,11 +1,11 @@
 module.exports = {
   stories: [
     '../stories/styles.js',
-    '../stories/components/*.js',
-    '../stories/elements/*.js',
-    '../stories/extensions/*.js',
-    '../stories/helpers/*.js',
-    '../stories/helpers/*.tsx',
+    '../stories/components/*.{js,tsx}',
+    '../stories/elements/*.{js,tsx}',
+    '../stories/extensions/*.{js,tsx}',
+    '../stories/helpers/*.{js,tsx}',
+    '../stories/helpers/*.{js,tsx}',
   ],
   addons: ['@storybook/preset-ie11', '@storybook/preset-scss'],
 }