+++ packages/dnb-eufemia/src/shared/index.js
@@ -8,5 +8,6 @@ import Provider from './Provider'
 import MediaQuery from './MediaQuery'
 import useMediaQuery from './useMediaQuery'
 export * from './SpacingHelper'
+export * from './interfaces'
 
 export { Context, Provider, MediaQuery, useMediaQuery }